Also add own user name to group chat name, open group chat if already exists
Own user name was not included in the group chat name, which would cause
collisions if two different users wanted to talk directly with the same
other group of people.

If a group chat with given name already exists, just open it instead of
reporting an error. This is way more user friendly.