Hack in nicer fake display names for group chats
Previous method of setting the group chat name once did not work for
our new dynamic approach, where we want to add users to the group chat
and have it seen, because it was cached somewhere and never updated.

This is an ugly hack that forces the group chat to be renamed each time
a member is added or removed, and parses the room name to fake a nicer
display.