Copy a configurable amount of chat history when creating new group chat
When creating a new group chat by adding users to a direct message room,
we want to copy some amount of history to give the new joiners some
context. This is currently limited to only copying from direct messages
into group chats, given the executing user belongs to both. The total
length of history is configurable in the settings, but should be kept low.